broke stake & mint integeration test into pieces
