+++ test_integration/04_redeem_and_unstake/01_single_staker_single_facilitator.test.ts
@@ -1,0 +1,15 @@
+import { execSync } from 'child_process';
+import * as path from 'path';
+const facilitatorKill = path.join(__dirname, '../kill_facilitator_process.sh');
+
+describe('redeem and unstake with single redeemer & facilitator process', async (): Promise<void> => {
+
+  before(async () => {
+
+  });
+
+  after(async (): Promise<void> => {
+    execSync(facilitatorKill, { stdio: [process.stdout, process.stderr], env: process.env });
+  });
+
+});
