+++ test_integration/facilitator_init.sh
