+++ test_integration/kill_facilitator_process.sh
